# doc/api.yml
openapi: 3.0.3
info:
  version: 'latest'
  title: KotiKota
servers:
  - url: 'http://192.168.0.245:8080'
security:
  - BearerAuth: [ ]
paths:
...
  /projects:
    get:
      tags:
        - Projects
      security: [ ]
      summary: Get the list of all projects
      operationId: getAllProjects
      parameters:
        - in: query
          name: ownerId
          description: Filter projects by ownerId
          schema:
            type: string
        - in: query
          name: page
          schema:
            $ref: '#/components/schemas/Page'
        - in: query
          name: pageSize
          schema:
            $ref: '#/components/schemas/PageSize'
      responses:
        200:
          description: List of existing projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
        400:
          $ref: "#/components/responses/400"
        500:
          $ref: "#/components/responses/500"
    put:
      tags:
        - Projects
      summary: Create or update projects
      operationId: crupdateProjects
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/CreateProject'
      responses:
        200:
          description: The created projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
        400:
          $ref: "#/components/responses/400"
        403:
          $ref: "#/components/responses/403"
        500:
          $ref: "#/components/responses/500"
        404:
          $ref: "#/components/responses/404"
  /projects/{projectId}:
    get:
      tags:
        - Projects
      summary: Get a specific project by id
      operationId: getProjectById
      security: [ ]
      parameters:
        - name: projectId
          in: path
          schema:
            type: string
          required: true
      responses:
        200:
          description: The Project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        500:
          $ref: "#/components/responses/500"
        404:
          $ref: "#/components/responses/404"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  responses:
    400:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BadRequestException'
    403:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/NotAuthorizedException'
    404:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ResourceNotFoundException'
...
    '500':
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/InternalServerException'
  schemas:
    Exception:
      type: object
      properties:
        type:
          type: string
        message:
          type: string
    InternalServerException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: InternalServerException
        message: Unexpected error
    BadRequestException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: BadRequestException
        message: Bad request
    NotAuthorizedException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: NotAuthorizedException
        message: Not authorized
    ResourceNotFoundException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: ResourceNotFoundException
        message: Resource of type <T> identified by <I> not found
...
    Page:
      type: integer
      description: Start at `1`
    PageSize:
      type: integer
      description: Cannot exceed `500`
...
    ProjectStatus:
      type: string
      enum:
        - OPEN
        - CLOSE
    ProjectHealth:
      type: string
      enum:
        - SUCCESS
        - FAILED
        - IN_PROGRESS
...
    CreateProject:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        targetAmount:
          type: integer
        deadline:
          type: string
          format: date-time
        ownerId:
          type: string
        status:
          $ref: '#/components/schemas/ProjectStatus'
        health:
          $ref: '#/components/schemas/ProjectHealth'
    Project:
      allOf:
        - $ref: "#/components/schemas/CreateProject"
      properties:
        totalDonations:
          type: integer
...